requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.6
django-dotenv==1.4.2
djangorestframework==3.8.2
drf-yasg==1.9.0
flex==6.13.2
future==0.16.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
kombu==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2-binary==2.7.5
pytz==2018.4
PyYAML==3.12
requests==2.19.1
rfc3987==1.3.7
ruamel.yaml==0.15.40
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.3.1
uritemplate==3.0.0
urllib3==1.23
validate-email==1.3
vine==1.1.4
